requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.1
django-appconf==1.0.6
django-compressor==4.4
django-cron==0.6.0
django-crontab==0.7.1
django-ipware==6.0.4
django-libsass==0.9
frozenlist==1.4.1
future==0.18.3
geocoder==1.38.1
geoip2==4.8.0
gunicorn==21.2.0
idna==3.6
kombu==5.3.5
libsass==0.23.0
maxminddb==2.5.2
multidict==6.0.5
netaddr==0.10.1
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
python-dateutil==2.8.2
python-ipware==2.0.1
ratelim==0.1.6
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
yookassa==3.0.0
